requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
astunparse=1.6.3=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.7.2=haa95532_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.3.2=pypi_0
flatbuffers=24.3.25=pypi_0
gast=0.4.0=pypi_0
google-auth=2.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.66.1=pypi_0
h5py=3.11.0=pypi_0
idna=3.10=pypi_0
joblib=1.4.2=pypi_0
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
libclang=18.1.1=pypi_0
libffi=3.4.4=hd77b12b_1
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
mdurl=0.1.2=pypi_0
ml-dtypes=0.4.1=pypi_0
namex=0.0.8=pypi_0
numpy=1.24.3=pypi_0
oauthlib=3.2.2=pypi_0
openssl=3.0.15=h827c3e9_0
opt-einsum=3.4.0=pypi_0
optree=0.12.1=pypi_0
packaging=24.1=pypi_0
pandas=2.2.3=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py310haa95532_0
protobuf=3.19.6=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pygments=2.18.0=pypi_0
python=3.10.14=he1021f5_1
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.2=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rich=13.8.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
setuptools=75.1.0=py310haa95532_0
six=1.16.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.10.0=pypi_0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-intel=2.17.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h0416ee5_0
typing-extensions=4.5.0=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
vc=14.40=h2eaa2aa_1
vs2015_runtime=14.40.33807=h98bb1dd_1
werkzeug=3.0.4=pypi_0
wheel=0.44.0=py310haa95532_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h8cc25b3_1
zlib=1.2.13=h8cc25b3_1
